Refactor PR reviewer into modular package structure
Extract PR reviewer functionality from monolithic script into organized package with separate modules for GitHub API, LLM analysis, data models, and core reviewer logic. This improves maintainability and code organization.
